+++ tests/test_bs4_cases.py
@@ -68,7 +68,7 @@ class SelectorNthOfTypeBugTest(unittest.TestCase):
     def test_parent_nth_of_type_preconditions(self):
         """Test `nth` type preconditions."""
 
-        els = sv.select('div > h1', self.soup, mode=sv.HTML)
+        els = sv.select('div > h1', self.soup, flags=sv.HTML)
         # check that there is a unique selection
         self.assertEqual(len(els), 1)
         self.assertEqual(els[0].string, 'An H1')
@@ -79,12 +79,12 @@ class SelectorNthOfTypeBugTest(unittest.TestCase):
         div_inner = h1.parent
         div_main = div_inner.parent
         div_main_children = [child for child in div_main.children]
-        self.assertEquals(div_main_children[0], '\n')
-        self.assertEquals(div_main_children[1], div_inner)
+        self.assertEqual(div_main_children[0], '\n')
+        self.assertEqual(div_main_children[1], div_inner)
 
     def test_parent_nth_of_type(self):
         """Test parent of `nth` of type."""
 
-        els = sv.select('div:nth-of-type(1) > h1', self.soup, mode=sv.HTML)
+        els = sv.select('div:nth-of-type(1) > h1', self.soup, flags=sv.HTML)
         self.assertEqual(len(els), 1)
         self.assertEqual(els[0].string, 'An H1')