+++ tests/test_soupsieve.py
@@ -33,10 +33,10 @@ class TestSoupSieve(unittest.TestCase):
         """
 
         soup = bs4.BeautifulSoup(markup, 'html5lib')
-        comments = [str(c).strip() for c in sv.comments(soup, mode=sv.HTML5)]
+        comments = [str(c).strip() for c in sv.comments(soup, flags=sv.HTML5)]
         self.assertEqual(sorted(comments), sorted(['before header', 'comment', "don't ignore"]))
 
-        comments = [str(c).strip() for c in sv.icomments(soup, limit=2, mode=sv.HTML5)]
+        comments = [str(c).strip() for c in sv.icomments(soup, limit=2, flags=sv.HTML5)]
         self.assertEqual(sorted(comments), sorted(['before header', 'comment']))
 
     def test_select(self):
@@ -71,6 +71,13 @@ class TestSoupSieve(unittest.TestCase):
             ids.append(el.attrs['id'])
 
         self.assertEqual(sorted(['5', 'some-id']), sorted(ids))
+
+        span = sv.select('span[id]', soup)[0]
+        ids = []
+        for el in sv.select('span[id]', span):
+            ids.append(el.attrs['id'])
+
+        self.assertEqual(sorted(['5']), sorted(ids))
 
     def test_match(self):
         """Test matching."""
@@ -122,22 +129,27 @@ class TestSoupSieve(unittest.TestCase):
         self.assertEqual(len(nodes), 1)
         self.assertEqual(nodes[0].attrs['id'], '6')
 
+        nodes = sv.filter('pre#6', [el for el in soup.html.body.children if isinstance(el, bs4.Tag)])
+        self.assertEqual(len(nodes), 1)
+        self.assertEqual(nodes[0].attrs['id'], '6')
+
     def test_copy_pickle(self):
         """Test copy and pickle."""
 
         # Test that we can pickle and unpickle
-        p1 = sv.compile('p[id]', mode=sv.HTML5)
+        p1 = sv.compile('p[id]', flags=sv.HTML5)
         sp1 = pickle.dumps(p1)
         pp1 = pickle.loads(sp1)
         self.assertTrue(pp1 == p1)
 
         # Test that we pull the same one from cache
-        p2 = sv.compile('p[id]', mode=sv.HTML5)
+        p2 = sv.compile('p[id]', flags=sv.HTML5)
         self.assertTrue(p1 is p2)
 
-        # Test that we compile a new one when providing a different mode
-        p3 = sv.compile('p[id]', mode=sv.HTML)
+        # Test that we compile a new one when providing a different flags
+        p3 = sv.compile('p[id]', flags=sv.HTML)
         self.assertTrue(p1 is not p3)
+        self.assertTrue(p1 != p3)
 
         # Test that the copy is equivalent, but not same.
         p4 = copy.copy(p1)
@@ -171,10 +183,36 @@ class TestSoupSieve(unittest.TestCase):
         self.assertTrue(p1 is p2)
 
         with pytest.raises(ValueError):
-            sv.compile(p1, mode=sv.HTML)
+            sv.compile(p1, flags=sv.HTML)
 
         with pytest.raises(ValueError):
             sv.compile(p1, namespaces={"": ""})
+
+    def test_immutable_object(self):
+        """Test immutable object."""
+
+        obj = sv.util.Immutable()
+
+        with self.assertRaises(AttributeError):
+            obj.member = 3
+
+    def test_immutable_dict(self):
+        """Test immutable dictionary."""
+
+        idict = sv.util.ImmutableDict({'a': 'b', 'c': 'd'})
+        self.assertEqual(2, len(idict))
+
+        with self.assertRaises(TypeError):
+            idict['a'] = 'f'
+
+        with self.assertRaises(TypeError):
+            sv.util.ImmutableDict([[3, {}]])
+
+        with self.assertRaises(TypeError):
+            sv.util.ImmutableDict([[{}, 3]])
+
+        with self.assertRaises(TypeError):
+            sv.ct.Namespaces({'a': {}})
 
 
 class TestDeprcations(unittest.TestCase):
@@ -263,3 +301,85 @@ class TestDeprcations(unittest.TestCase):
             self.assertEqual(sorted(comments), sorted(['before header', 'comment']))
             self.assertTrue(len(w) == 1)
             self.assertTrue(issubclass(w[-1].category, DeprecationWarning))
+
+
+class TestInvalid(unittest.TestCase):
+    """Test invalid."""
+
+    def test_invalid_mode(self):
+        """Test invalid mode."""
+
+        with self.assertRaises(ValueError):
+            sv.compile('p', None, sv.util.HTML | sv.util.HTML5)
+
+    def test_invalid_combination(self):
+        """
+        Test invalid combination.
+
+        Selectors cannot start with relational symbols unless in `:has()`.
+        `:has()` cannot start with `,`.
+        """
+
+        with self.assertRaises(SyntaxError):
+            sv.compile('> p')
+
+        with self.assertRaises(SyntaxError):
+            sv.compile(', p')
+
+        with self.assertRaises(SyntaxError):
+            sv.compile(':has(, p)')
+
+        with self.assertRaises(SyntaxError):
+            sv.compile('div >> p')
+
+        with self.assertRaises(SyntaxError):
+            sv.compile('div >')
+
+    def test_invalid_pseudo_close(self):
+        """Test invalid pseudo close."""
+
+        with self.assertRaises(SyntaxError):
+            sv.compile('div)')
+
+        with self.assertRaises(SyntaxError):
+            sv.compile(':is(div,)')
+
+    def test_invalid_pseudo_open(self):
+        """Test invalid pseudo close."""
+
+        with self.assertRaises(SyntaxError):
+            sv.compile(':is(div')
+
+    def test_invalid_incomplete_has(self):
+        """Test invalid `:has()`."""
+
+        with self.assertRaises(SyntaxError):
+            sv.compile(':has(>)')
+
+        with self.assertRaises(SyntaxError):
+            sv.compile(':has()')
+
+    def test_invalid_tag(self):
+        """
+        Test invalid tag.
+
+        Tag must come first.
+        """
+
+        with self.assertRaises(SyntaxError):
+            sv.compile(':is(div)p')
+
+    def test_invalid_syntax(self):
+        """Test invalid syntax."""
+
+        with self.assertRaises(SyntaxError):
+            sv.compile('div?')
+
+    def test_invalid_namespace(self):
+        """Test invalid namespace."""
+
+        with self.assertRaises(TypeError):
+            sv.ct.Namespaces(((3, 3),))
+
+        with self.assertRaises(TypeError):
+            sv.ct.Namespaces({'a': {}})