+++ docs/src/markdown/about/changelog.md
@@ -1,5 +1,10 @@
 # Changelog
 
+## 0.6.0
+
+- **NEW**: `mode` attribute is now called `flags` to allow for other options in the future.
+- **FIX**: More corner cases for `nth` selectors.
+
 ## 0.5.3
 
 - **FIX**: Previously, all pseudo classes' selector lists were evaluated as one big group, but now each pseudo classes' selector lists are evaluated separately.